fix: all sections with their buttons
