+++ src/modules/profile/public/PublicProfile.tsx
@@ -1,14 +1,14 @@
-import { Link } from 'react-router-dom';
+import { useState } from 'react';
 
 import FacebookIcon from '@mui/icons-material/Facebook';
 import LinkedInIcon from '@mui/icons-material/LinkedIn';
 import TwitterIcon from '@mui/icons-material/Twitter';
-import { Button, Stack, Typography } from '@mui/material';
+import { Button, Typography } from '@mui/material';
 
 import SocialLinks from 'social-links';
 
+import BorderedSection from '@/components/layout/BorderedSection';
 import { useAccountTranslation } from '@/config/i18n';
-import { PUBLIC_PROFILE_PATH } from '@/config/paths';
 import { hooks } from '@/config/queryClient';
 import {
   PUBLIC_PROFILE_BIO_ID,
@@ -21,8 +21,8 @@ import {
   PUBLIC_PROFILE_TWITTER_ID,
 } from '@/config/selectors';
 
-import RoundedStack from '../common/RoundedStack';
-import DisplayingMemberPublicProfileLinks from './DisplayingMemberPublicProfileLinks';
+import DisplayLinks from './DisplayLinks';
+import EditPublicProfile from './EditPublicProfile';
 
 const MemberPublicProfile = (): JSX.Element => {
   const socialLinks = new SocialLinks();
@@ -32,20 +32,29 @@ const MemberPublicProfile = (): JSX.Element => {
 
   const { bio, linkedinID, twitterID, facebookID } = publicProfile || {};
 
+  const [isEditing, setIsEditing] = useState(false);
+
+  const onClose = () => setIsEditing(false);
+  const onOpen = () => setIsEditing(true);
+
+  if (isEditing) {
+    return <EditPublicProfile onClose={onClose} />;
+  }
   return (
-    <RoundedStack>
-      <Stack direction="row" justifyContent="space-between">
-        <Typography variant="h5">{t('PUBLIC_PROFILE_TITLE')}</Typography>
+    <BorderedSection
+      title={t('PUBLIC_PROFILE_TITLE')}
+      topActions={[
         <Button
-          component={Link}
-          to={PUBLIC_PROFILE_PATH}
+          key="edit"
           variant="contained"
+          onClick={onOpen}
           id={PUBLIC_PROFILE_EDIT_BUTTON_ID}
           size="small"
         >
           {t('EDIT_BUTTON_LABEL')}
-        </Button>
-      </Stack>
+        </Button>,
+      ]}
+    >
       <Typography variant="body1" color="textSecondary">
         {t('PUBLIC_PROFILE_BIO')}
       </Typography>
@@ -53,7 +62,7 @@ const MemberPublicProfile = (): JSX.Element => {
         {bio || t('PUBLIC_PROFILE_BIO_EMPTY_MSG')}
       </Typography>
       {linkedinID ? (
-        <DisplayingMemberPublicProfileLinks
+        <DisplayLinks
           icon={<LinkedInIcon />}
           contentId={PUBLIC_PROFILE_LINKEDIN_ID}
           href={socialLinks.sanitize('linkedin', linkedinID)}
@@ -61,7 +70,7 @@ const MemberPublicProfile = (): JSX.Element => {
           hrefId={PUBLIC_PROFILE_LINKEDIN_HREF_ID}
         />
       ) : (
-        <DisplayingMemberPublicProfileLinks
+        <DisplayLinks
           icon={<LinkedInIcon />}
           contentId={PUBLIC_PROFILE_LINKEDIN_ID}
           content={t('PUBLIC_PROFILE_LINKEDIN_EMPTY_MSG')}
@@ -69,7 +78,7 @@ const MemberPublicProfile = (): JSX.Element => {
       )}
 
       {twitterID ? (
-        <DisplayingMemberPublicProfileLinks
+        <DisplayLinks
           icon={<TwitterIcon />}
           contentId={PUBLIC_PROFILE_TWITTER_ID}
           href={socialLinks.sanitize('twitter', twitterID)}
@@ -77,14 +86,14 @@ const MemberPublicProfile = (): JSX.Element => {
           hrefId={PUBLIC_PROFILE_TWITTER_HREF_ID}
         />
       ) : (
-        <DisplayingMemberPublicProfileLinks
+        <DisplayLinks
           icon={<TwitterIcon />}
           contentId={PUBLIC_PROFILE_TWITTER_ID}
           content={t('PUBLIC_PROFILE_TWITTER_EMPTY_MSG')}
         />
       )}
       {facebookID ? (
-        <DisplayingMemberPublicProfileLinks
+        <DisplayLinks
           icon={<FacebookIcon />}
           contentId={PUBLIC_PROFILE_FACEBOOK_ID}
           href={socialLinks.sanitize('facebook', facebookID)}
@@ -92,13 +101,13 @@ const MemberPublicProfile = (): JSX.Element => {
           hrefId={PUBLIC_PROFILE_FACEBOOK_HREF_ID}
         />
       ) : (
-        <DisplayingMemberPublicProfileLinks
+        <DisplayLinks
           icon={<FacebookIcon />}
           contentId={PUBLIC_PROFILE_FACEBOOK_ID}
           content={t('PUBLIC_PROFILE_FACEBOOK_EMPTY_MSG')}
         />
       )}
-    </RoundedStack>
+    </BorderedSection>
   );
 };
